elf_reader: add tests for ELF objects without BTF
Adds infrastructure to test loading of ELF objects compiled without BTF information.
